Please add an Note in the line 80
Please add an Note in the line 80 to explain the customer why is so important to purchase an Intune license if they have an Android or iOS devices managed by Office 365 MDM and they are about to use co-management with a Win10 devices:

Since co-management with Intune is an Intune feature, enabling it for a user will exclude that user from MIFO (Office MDM) per above, so the customer should be careful deciding which users are intended to have co-managed PCs. If those users have other devices currently managed by MIFO (e.g. iOS, Android, or Windows PCs not in SCCM), the customer will need to purchase additional Intune licenses to manage those with Intune, as co-management does not cover managing such devices, and enabling co-management for that user will disable MIFO.