Please add a note to remark the Azure Environment.
Please add a note to remark on the importance of picking the correct Azure Environment for Government customers.
Azure US Government Cloud is for GCC-H environments only. I had a Gov case (19864207 ) where the customer was struggling for weeks trying to set up co-management.

He was getting: "Access Denied - You are not authorized to view this page. Your IT department may be able to help" after every sign in attempts and we confirmed he was trying to authenticate against the wrong (.us) Azure Endpoints instead .com.

This is reproducible and if you have any questions or need more clarification you can ping me mibulgad@microsoft.com